+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/screens/settings/SettingsUiState.kt
@@ -1,3 +1,5 @@
 package net.nymtech.nymvpn.ui.screens.settings
 
-data class SettingsUiState(val daemonVersion: String = "", val isMixnetTuningEnabled: Boolean = false)
+import net.nymtech.nymvpn.ui.screens.settings.components.SubscriptionUiState
+
+data class SettingsUiState(val daemonVersion: String = "", val isMixnetTuningEnabled: Boolean = false, val subscription: SubscriptionUiState? = null)