+++ nym-vpn-android/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Add Subscription and Bandwidth information (https://github.com/nymtech/nym-vpn-client/pull/4817)
 
 ### Changed
 - AGP version update (https://github.com/nymtech/nym-vpn-client/pull/4586)